requirements.txt
# Core functionality
langchain>=0.1.0
langchain-huggingface>=0.0.5
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0

# Web interface
flask>=2.0.0

# Data processing
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0